requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
psycopg2==2.9.9
